+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "static",
   "version": "1.0.0",
-  "description": "",
+  "description": "Revised a portfolio with flask",
   "main": "index.js",
   "directories": {
     "lib": "lib"
@@ -13,6 +13,7 @@
   "license": "ISC",
   "devDependencies": {
     "autoprefixer": "^9.7.6",
+    "browser-sync": "^2.26.7",
     "gulp": "^4.0.2",
     "gulp-notify": "^3.2.0",
     "gulp-plumber": "^1.2.1",
@@ -24,5 +25,13 @@
     "last 2 version",
     "> 1%",
     "IE 11"
-  ]
+  ],
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/mhiratsuka/flask-portfolio.git"
+  },
+  "bugs": {
+    "url": "https://github.com/mhiratsuka/flask-portfolio/issues"
+  },
+  "homepage": "https://github.com/mhiratsuka/flask-portfolio#readme"
 }
